Split the command definition source file for WASM
