+++ pkg/cmd/commands/server/ssh_parameter.go
@@ -1,0 +1,37 @@
+// Copyright 2017-2020 The Usacloud Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package server
+
+import (
+	"github.com/sacloud/usacloud/pkg/cmd/cflag"
+)
+
+type sshParameter struct {
+	cflag.ZoneParameter   `cli:",squash" mapconv:",squash"`
+	cflag.IDParameter     `cli:",squash" mapconv:",squash"`
+	cflag.CommonParameter `cli:",squash" mapconv:"-"`
+
+	Key            string `cli:",short=i" validate:"omitempty,file"`
+	User           string `cli:",short=l"`
+	Port           int    `cli:",short=p" validate:"required"`
+	Password       string `cli:",aliases=pass-phrase"`
+	WaitUntilReady bool   `cli:",aliases=wait"`
+}
+
+func newSSHParameter() *sshParameter {
+	return &sshParameter{
+		Port: 22,
+	}
+}
